backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.3
psycopg2-binary==2.9.7
Flask-Migrate==4.0.4
python-dotenv==1.0.0
Flask-CORS==4.0.1
Flask-Mail==0.9.1
